fix(room-session): address review comments
