+++ infrastructure/.env.example
@@ -3,6 +3,7 @@ LIVEKIT_API_SECRET=devsecret
 LIVEKIT_WS_URL=ws://localhost:7880
 AUTH_BIND_ADDR=0.0.0.0:8787
 AUTH_DATABASE_URL=sqlite:///app/data/auth.db?mode=rwc
+# Dev-only placeholder. In staging/prod, use a long random secret (32+ bytes) from your secret manager.
 AUTH_COOKIE_SECRET=dev-cookie-secret-change-me
 AUTH_SESSION_TTL_SECONDS=604800
 AUTH_BOOTSTRAP_ADMIN_EMAILS=
